% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_ant}
\alias{process_ant}
\title{Process Attention Network Test experiment data file}
\usage{
process_ant(ant_file, p, time, json = TRUE)
}
\arguments{
\item{ant_file}{Path to ANT file}

\item{p}{Participant number}

\item{time}{Time-point at which participant completed ANT}

\item{json=TRUE}{Boolean indicating whether ANT data format is JSON (TRUE) or CSV (FALSE)}
}
\value{
Data frame
}
\description{
\code{process_ant()} processes data files generated by the Attention Network Task experiment
(\url{https://expfactory.github.io/attention_network_task.html}).
}
\examples{
ant <- expand.grid(token = participants$token, time=1:2) \%>\%
 rowwise() \%>\%
 mutate(path=paste('data/', token, '_finished/ant', time, '-results.json', sep='')) \%>\%
 do(., efantr::process_ant(.$path, .$token, .$time)) \%>\%

}
\keyword{ANT}
\keyword{expfactory}
